+++ idea/src/org/jetbrains/kotlin/idea/refactoring/safeDelete/KotlinSafeDeleteProcessor.kt
@@ -185,6 +185,10 @@ class KotlinSafeDeleteProcessor : JavaSafeDeleteProcessor() {
 
         return when (element) {
             is KtClassOrObject -> {
+                if (element is KtEnumEntry) {
+                    LightClassUtil.getLightClassBackingField(element)?.let { findUsagesByJavaProcessor(it, false) }
+                }
+
                 element.toLightClass()?.let { klass ->
                     findDelegationCallUsages(klass)
                     findUsagesByJavaProcessor(klass, false)